pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bus-searcher"
version = "0.1.0"
description = "Bus ridership statistics analysis system for Pangyo-dong, Seongnam, Gyeonggi"
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = "MIT"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic>=2.5.0,<3.0.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.17.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "APScheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "databases[postgresql]>=0.6.0",  # Added database library
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
